influxd: Support backing up to stdout
Generates a tar that contains database metadata and all shards. The
shards in the tar contain the tsm files directly (as opposed to regular
backups that have those in a separate tar), which both allows creating
the backup in streaming fashion and removes the need to separately
extract the tsm files while restoring backup. Like with the -portable,
switch, when not explicitly specifying the database to back up, all
databases are backed up.