Support restoring backup directly from stdin
Expects a tar that is either manually created from output directory
of a regular backup or created with `backup -write-to-stdout`.

Note that this code assumes database names do not contain . or /.
Restoring databases that do have those characters in their names does
not work with the -read-from-stdin option.

To make backups created with -write-to-stdout work with regular
(non-stdin) restoration, update that logic to handle tsm files that
are not wrapped inside a tar but reside on the disk directly and no
tar packages on disk as well as meta file named just "meta" instead of
meta.xx.