Fix restoring shards that end in zeros
When creating backups the shard id is converted into integer and then
back to string using %05d formatting. When restoring there was code that
trimmed zeros from the shard identifier. However, trimming was done both
from beginning and end of the string so shard id 00100 became 1 instead
of 100 as it presumably should've. This essentially dropped the shard
when restoring and subsequent backups no longer contained it.

This would seem to be a bug in InfluxDB itself and the same bug was
carried over to the streaming restoration. Fix all trims to only remove
zeros on the left.