backup: Allow ignoring errors caused by missing shard
A shard might go away after server has returned list of shards. Trying
to back up the shard always results in empty response from server, which
previously was considered to be fatal error (unless all errors were
being ignored). This should be harmless case where we can just log what
happened and proceed with the next shard.

As the server protocol does not allow providing any extra information
regarding the failure (server either returns actual file data or not)
the check needs to be done on client side and as the client does not
know where data resides (and in some cases client might have different
view of the filesystem), allow passing -data-dir parameter identifying
the location of the shards so that client can produce full path for the
existence check.